refactor: change get_default_csv_data() to return panda series and fixed test_map_xds_to_healthri_dcat_dataset_returns_model
